backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0

